Add antd, antd-css-utilities, chart.js, clsx, react-chartjs-2 dependencies and update Home component to remove usage of styles.
This commit adds the following dependencies: antd, antd-css-utilities, chart.js, clsx, and react-chartjs-2. It also updates the Home component to remove the usage of styles, including the import statement for styles, the className attributes in the Image components, and the div with the styles.center class. These changes are part of updating the project to use the new app directory structure in Next.js 14.1.4.